+++ api/src/utils/domain/business-rule-validation.exception.ts
